move all to components folder and remove unused icons
