+++ apps/lite/ui/src/Updater.tsx
@@ -2,8 +2,8 @@ import { AlertDialog } from "@base-ui/react";
 import { useEffect, useState } from "react";
 import type { FC } from "react";
 import { Button } from "#ui/components/Button.tsx";
-import { classes } from "#ui/ui/classes.ts";
-import uiStyles from "#ui/ui/ui.module.css";
+import { classes } from "#ui/components/classes.ts";
+import uiStyles from "#ui/components/ui.module.css";
 
 export const Updater: FC = () => {
 	const [version, setVersion] = useState<string | null>(null);
