+++ apps/lite/ui/src/components/Button.tsx
@@ -1,5 +1,5 @@
 import styles from "./Button.module.css";
-import { classes } from "#ui/ui/classes.ts";
+import { classes } from "#ui/components/classes.ts";
 import { ComponentProps, ReactNode, forwardRef, isValidElement } from "react";
 
 type ButtonVariant = "pop" | "gray" | "outline" | "danger" | "ghost" | "inverted";
