+++ apps/lite/ui/src/components/Spinner.tsx
@@ -1,5 +1,5 @@
 import styles from "./Spinner.module.css";
-import { classes } from "#ui/ui/classes.ts";
+import { classes } from "#ui/components/classes.ts";
 import { ComponentProps, FC } from "react";
 
 export const Spinner: FC<ComponentProps<"span">> = ({ className, ...props }) => (
