+++ apps/lite/ui/src/routes/project/$id/workspace/DependencyIndicatorButton.tsx
@@ -2,8 +2,8 @@ import { headInfoQueryOptions } from "#ui/api/queries.ts";
 import { getBranchNameByCommitId } from "#ui/api/ref-info.ts";
 import { projectActions } from "#ui/projects/state.ts";
 import { useAppDispatch } from "#ui/store.ts";
-import { classes } from "#ui/ui/classes.ts";
-import uiStyles from "#ui/ui/ui.module.css";
+import { classes } from "#ui/components/classes.ts";
+import uiStyles from "#ui/components/ui.module.css";
 import { Tooltip, useRender } from "@base-ui/react";
 import { useQuery } from "@tanstack/react-query";
 import { Array, pipe } from "effect";
