+++ apps/lite/ui/src/routes/project/$id/workspace/OperationTarget.tsx
@@ -8,7 +8,7 @@ import {
 	type OperationType,
 	useRunOperation,
 } from "#ui/operations/operation.ts";
-import { classes } from "#ui/ui/classes.ts";
+import { classes } from "#ui/components/classes.ts";
 import { projectActions, selectProjectOutlineModeState } from "#ui/projects/state.ts";
 import { useAppDispatch, useAppSelector } from "#ui/store.ts";
 import { dropTargetForElements } from "@atlaskit/pragmatic-drag-and-drop/element/adapter";
